Debugging

+++ src/main/java/edu/uark/registerapp/models/entities/ProductEntity.java
@@ -53,13 +53,13 @@ public class ProductEntity {
 	}
 
     @Column(name="price")
-    private long price;
+    private float price;
 
-	public long getPrice() {
+	public float getPrice() {
 		return this.price;
 	}
 
-	public ProductEntity setPrice(final long price) {
+	public ProductEntity setPrice(final float price) {
 		this.price = price;
 		return this;
 	}
@@ -84,12 +84,12 @@ public class ProductEntity {
 
 	public ProductEntity() {
 		this.count = -1;
-		this.price = 0L;
+		this.price = 0.00;
 		this.id = new UUID(0, 0);
 		this.lookupCode = StringUtils.EMPTY;
 	}
 
-	public ProductEntity(final String lookupCode, final int count, final long price) {
+	public ProductEntity(final String lookupCode, final int count, final float price) {
 		this.count = count;
 		this.price = price;
 		this.id = new UUID(0, 0);
@@ -97,7 +97,7 @@ public class ProductEntity {
 	}
 
 	public ProductEntity(final Product apiProduct) {
-		this.price = 0L;
+		this.price = 0.00;
 		this.id = new UUID(0, 0);
 		this.count = apiProduct.getCount();
 		this.lookupCode = apiProduct.getLookupCode();
